+++ src/main/java/com/ahenkeshi/unikko/Unikko.java
@@ -1,7 +1,11 @@
+/* This file is a part of Unikko Utility Mod: https://github.com/jnkyto/Unikko which is
+distributed under CC0-1.0: https://creativecommons.org/publicdomain/zero/1.0/legalcode
+*/
+
 package com.ahenkeshi.unikko;
 
 import com.ahenkeshi.unikko.cmd.CommandHandler;
-import com.ahenkeshi.unikko.utils.config.SoftConfigUtils;
+import com.ahenkeshi.unikko.utils.config.SoftConfig;
 import com.ahenkeshi.unikko.utils.config.HardConfigUtils;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.ModInitializer;
@@ -17,7 +21,7 @@ public class Unikko implements ClientModInitializer, ModInitializer {
 
 	public static final String MODID = "Unikko";
 	public static final String VERSION = "1.1.0";
-	public static String REL_DATE = "2021/10/09";
+	public static String REL_DATE = "2021/11/18";
 	public static String DEV = "-dev";
 
 	public static final ModContainer MOD = FabricLoader.getInstance()
@@ -25,6 +29,7 @@ public class Unikko implements ClientModInitializer, ModInitializer {
 			.orElseThrow(NullPointerException::new);
 
 	public static final Logger logger = LogManager.getLogger();
+	public static SoftConfig softConfig;
 
 	@Override
 	public void onInitialize()	{
@@ -32,7 +37,6 @@ public class Unikko implements ClientModInitializer, ModInitializer {
 		long initStart = System.currentTimeMillis();
 
 		try {HardConfigUtils.createFile();} catch (IOException e) {e.printStackTrace();}
-		SoftConfigUtils.saveBooleansToConfigFile();
 
 		if(!DEV_MODE){
 			DEV = "";
@@ -41,6 +45,7 @@ public class Unikko implements ClientModInitializer, ModInitializer {
 
 		CommandHandler.init();
 		long initDuration = System.currentTimeMillis() - initStart;
+		softConfig = new SoftConfig(HardConfigUtils.getConfigMap());
 		logger.info("Initialization took " + initDuration + "ms.");
 	}