+++ src/main/java/com/ahenkeshi/unikko/mixin/MixinTitleScreen.java
@@ -4,8 +4,8 @@ distributed under CC0-1.0: https://creativecommons.org/publicdomain/zero/1.0/leg
 
 package com.ahenkeshi.unikko.mixin;
 
+import com.ahenkeshi.unikko.Unikko;
 import com.ahenkeshi.unikko.modules.DiscordRPC;
-import com.ahenkeshi.unikko.utils.config.SoftConfigUtils;
 import net.minecraft.client.gui.screen.TitleScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -16,7 +16,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class MixinTitleScreen {
     @Inject(at = @At("HEAD"), method = "init()V")
     private void init(CallbackInfo info)    {
-        SoftConfigUtils.saveBooleansToConfigFile();
+        Unikko.softConfig.pushHard();
         DiscordRPC.init();
     }
 }